refactor(S-3.02): extract handle_tickets to assets/tickets.rs
Moves handle_tickets + filter_tickets (with --open colorName filter)
+ all filter_tickets unit tests to tickets.rs shard. Tests co-located
with the function they exercise per Rust convention. All 612 unit tests
pass.
